requirements.txt
tensorflow-gpu
opencv-python
Pillow